New command line option: -kE
Check the encoding form used for those attributes that use constant values
and determine if the number of bytes used to represent the attribute data
can be reduced by using unsigned LEB128 encoding. Gives a table with the
current usage, the usage if LEB128 is used and a percentage rate.

Signed-off-by: Carlos Alberto Enciso <carlos.alberto.enciso@gmail.com>