Summary: Fixes some make time warnings by doing the build-time builds the automake way.
	modified:   README
noted the example install commands for libelf and zlib
work on 18.04 Ubuntu as well as 16.04.

    configure.ac: Updated version to 20180901.
Corrected a couple of typos in the summary output of configure.
    configure: Regenerated.

	modified:   dwarfdump/Makefile.am
Changed the way to build tag_tree and tag_attr
executables to the automake way, eliminating make warnings
when building.

	modified:   libdwarf/Makefile.am
Changed the way to build gennames
executable to the automake way, eliminating make warnings
when building.