Removed an unneeded struct esb_s struct. It was missing its destructor call so it actually leaked some memory. 	modified:   dwarfdump/print_pubnames.c
Made general.h idempotent with #ifdef GENERAL_H
	modified:   dwarfgen/general.h
Fixed bug in IRAttr assignment constructor.
	modified:   dwarfgen/irepdie.h
Added dropFormData member function to IRAttr
though it is not strictly necessary.
Makes it perhaps a bit clearer that the form
information is stale at certain points.
	modified:   dwarfgen/ireptodbg.cc

Now when there are zero bytes to copy we do not call memcpy,
as in that case the source pointer will be NULL.
The issue was reported by gcc with -fsanitize.
	modified:   libdwarf/pro_section.c