Using cmake to copy C code and modify it fails dramatically as cmake and the ; character have a tight and inflexible relationship.
So now we create an extra copy assuming _Elf.
   libdwarf/generated_libdwarf.h.in
cmake and configure simply copy
or the other *libdwarf.h.in to libdwarf.h in
the build area. Both can do that safely.

scripts/buildstandardsource.sh is run
by a developer (and requires little more
than sh, sed and compiling
and running simple C code) so
it is easily dealt with.

	modified:   configure.cmake
	new file:   libdwarf/generated_libdwarf.h.in
	modified:   scripts/buildstandardsource.sh
	modified:   scripts/fixlibdwarfelf.sh