The latest is in ChangeLog. Renamed the 2019 	modified:   dwarfdump/ChangeLog 	copied:     dwarfdump/ChangeLog -> dwarfdump/ChangeLog2019
Bugs noted by Coverity Scan. These bugs would not crash
anything, but could lead to a leak in case of certain
unlikely errors.
	modified:   dwarfdump/dwconf.c
	modified:   dwarfdump/print_die.c

This was not properly treating % characters in
strings it sanitizes: a few lines of code were
too far down in a couple of routines.
	modified:   dwarfdump/sanitized.c
`
The latest is in ChangeLog
	modified:   libdwarf/ChangeLog
	modified:   libdwarf/dwarf_query.c

The changelogs have additional details on the errors.