Delete all the tsearch/* directory stuff, that is now on github as the project davea42 tsearch-code 	modified:   Makefile.am
Remove UNUSEDARG and delete an unused
libdwarf-internal argument.
	modified:   src/lib/libdwarf/dwarf_fission_to_cu.c

Add site-specific text to the error message that is generated
when an allocation fails.
	modified:   src/lib/libdwarf/dwarf_gdbindex.c

Add include of libdwarf_private.h so it builds again.
It is not part of libdwarf, it is a support program
for libdwarf.
	modified:   src/lib/libdwarf/gennames.c

Correct the references in MAINTAINERCLEANFILES
and format part of the table to look nicer.
	modified:   test/Makefile.am