Neither libdwarf nor libdwarfp ever abort() now.
In case there is no way to report to
the caller what the error is (because
the caller provided no means to report error
messages) libdwarf prints a one-line
message to stderr with the error number.
For errors libdwarf and libdwarfp always
return DW_DLV_ERROR.

	modified:   doc/libdwarf.mm
	modified:   doc/libdwarf.pdf
	modified:   src/lib/libdwarf/NEWS
	modified:   src/lib/libdwarf/dwarf_error.c
	modified:   src/lib/libdwarfp/pro_error.c

Sligtly improved the message from simplereader in case of a
libdwarf-detected error.
	modified:   src/bin/dwarfexample/simplereader.c