if 0 the glflags.gf_debug_names_flag code as it is not correct yet. 	modified:   src/bin/dwarfdump/dwarfdump.c A beginning of some revisions. Ignore this for now. 	modified:   src/bin/dwarfdump/print_debug_names.c
Correcting the use  of dwarf_global_formref()
where the reference can be DW_AT_type and DW_FORM_ref_sig8
so we understand what section is being targeted by
an offset. And can check and report the target die TAG
and name (if it has a DW_AT_name).
	modified:   src/bin/dwarfdump/print_die.c

Just delete a pointless blank line.
	modified:   src/bin/dwarfdump/print_pubnames.c