These are no longer #include-d anywhere. The content is always in  libdwarf.h         deleted:    src/bin/dwarfdump/dwarf_names.h         deleted:    src/lib/libdwarf/dwarf_names.h
Removed all references to dwarf_names.h except
in a single CLEANFILES
Ensured that the compiler warning options are passed
to all the test files (there were typos in that Makefile.am)
        modified:   src/bin/buildopstab/Makefile.am
        modified:   src/bin/dwarfdump/CMakeLists.txt
        modified:   src/bin/dwarfdump/Makefile.am
        modified:   src/bin/gennames/Makefile.am
        modified:   src/lib/libdwarf/Makefile.am
        modified:   test/Makefile.am
        modified:   src/lib/libdwarf/CMakeLists.txt

Removed obsolete options and unused variables.
We sort the output function names dwarf_names.h by character
compares, where lowercase is elevated to uppercase
for the comparison, which
achieves a better look in the output.
The dwarf_names.h is a candidate to replace the
list of these protypes in libdwarf.h
        modified:   src/bin/gennames/gennames.c

Removed unused argument. Where Argument required
but unused (when faking something in libdwarf
so the test program will link) marked as UNUSEDARG to silence
warning.
        modified:   test/test_linkedtopath.c