If shared-library build requested no archive library is built (used to build both on request).         modified:   CMakeLists.txt
Alterations to build opscounttab.c properly.
	modified:   src/bin/buildopstab/Makefile.am

Now does more checking and adds a useful
comment in the output opscounttab.c where
we have a DW_OP with two spellings (which are extensions,
not standard DW_OP instances).
Reads dwarf.h to be sure the fast-access table
has all the named DW_OP in dwarf.h present in the
generated source..
Now includes config.h.
data is static not malloc-d.
	modified:   src/bin/buildopstab/buildopscounttab.c

Now with the better messages.
	modified:   src/bin/dwarfdump/opscounttab.c

The externally-visible table is now _dwarf_opscounttab
as it is not supposed to be visible or used outside the library.
	modified:   src/bin/dwarfdump/opscounttab.h

dwarf_opscounttab->_dwarf_opscounttab
	modified:   src/bin/dwarfdump/print_die.c