A recent update changed things.  make dist was now causing test/dummyexecutable to be rebuilt from test/dummyexecutable.c.  That was not intended. So renamed dummyexecutable.c so make cannot make the connection or try to rebuild dummyexecutable.
This simple change just makes what is going on slightly
clearer with respect to the two object sections involved.
	modified:   src/lib/libdwarf/dwarf_debuglink.c

This file should not be in libdwarfp, libdwarfp gets the
content from libdwarf source.
	deleted:    src/lib/libdwarfp/dwarf_string.c

Added an exit 0 so even if run the script does nothing. As intended!
It is really documentation and should not be run.
	modified:   test/buildingdummy.sh
	renamed:    test/dummyexecutable.c -> test/dummysourceignore

The makefile reflects the name changes so the files get into
the make dist output.
	modified:   test/Makefile.am