Now, for the first time, DW_LIBDWARF_VERSION_MACRO DW_LIBDWARF_VERSION_MINOR DW_LIBDWARF_VERSION_MICRO and DW_LIBDWARF_VERSION are #defined. These have the Semantic version number. At present Macro is 0 minor is 3 Micro is 0 DW_LIBDWARF_VERSION is defined as "0.3.0"
These are the version defined in libdwarf.h
when one includes it.
No earlier libdwarf.h defined any of these.

The function dwarf_package_version(),
on the other hand, returns a string
with the PACKAGE_VERSION visible
in configure.ac and CMakeLists.txt
at the time the library was compiled

	modified:   src/lib/libdwarf/libdwarf.h