Now builds a new opscounttab which has the change in the size of the expression stack for each DW_OP. To enable the print_die.c changes.     modified:   src/bin/buildopstab/buildopscounttab.c     modified:   src/bin/dwarfdump/opscounttab.c (regenerated)     modified:   src/bin/dwarfdump/opscounttab.h (new field added)
Unless the expression is trivial
printing expressions (DW_OP_swap etc)
-v -i now adds a line like
  DW_OPs= 13 maxstackdepth= 4 endingstackdepth= 3
at the end of the list of DW_OP expression operations
as a shorthand for the instructions and their
use of an expression stack.
If there are expression loops It adds
  loopcount= <n>
to the end of the line (counting the
instances of a DW_OP_bra or DW_OP_skip
in a negative direction).
    modified:   src/bin/dwarfdump/print_die.c

The old names of these functions were a bit
misleading. They are for all expressions.
dwarfdump_print_location_operations ->
  dwarfdump_print_expression_operations
print_location_operations ->
  print_expression_operations
    modified:   src/bin/dwarfdump/esb_using_functions.h
    modified:   src/bin/dwarfdump/print_frames.c
    modified:   src/bin/dwarfdump/print_frames.h
    modified:   src/bin/dwarfdump/print_hipc_lopc_attr.c
    modified:   src/bin/dwarfdump/print_tag_attributes_usage.c