To avoid warnings from a new compiler (warnings seen now for the first time) we remove some C with undefined meaning such as
headdepth = head->llink - (struct ts_entry *)0;

See ISO C 1990 section 6.3.6.

We add a specific field to count the balanced tree depth instead
of using strange pointer arithmetic on llink.
No need for pointer arithmetic at all.
There is a better way to do this using a HEAD
separate struct and a pointer-to an rlink field
as that would reduce struct size here,
but we are postponing that for now.
	modified:   src/bin/dwarfdump/dd_tsearchbal.c