Added include libdwarf_private.h as an aid to testing behaviour when malloc space exhausted (Faking exhaustion via a tiny amount of code NOT in the source tree). 	modified:   src/bin/dwarfdump/dd_addrmap.c 	modified:   src/bin/dwarfdump/dd_command_options.c 	modified:   src/bin/dwarfdump/dd_compiler_info.c 	modified:   src/bin/dwarfdump/dd_esb.c 	modified:   src/bin/dwarfdump/dd_helpertree.c 	modified:   src/bin/dwarfdump/dwarfdump.c
Added a fixed=size static array (with 5 entries) so in
the very rare case of
an error not attributable to a Dwarf_Debug
it is recorded and can be freed later.
Caused by memory exhaustion (alloc fail)
or by an object file with a very serious corruption
making it impossible to open in libdwarf.
See _dwarf_flush_static_error_list()
	modified:   src/lib/libdwarf/dwarf_alloc.c
Adding the static_err list function declarations (two of them)
	modified:   src/lib/libdwarf/dwarf_alloc.h

Initializing a local pointer to 0. Fixing DEBUG
messages that had copy-paste errors.
	modified:   src/lib/libdwarf/dwarf_error.c

Added call to _dwarf_flush_static_error_list()
	modified:   src/lib/libdwarf/dwarf_init_finish.c