_dwarf_memcpy_swap_bytes is now simpler and accomodates all words 8 bytes or less long by swapping bytes.   modified:   src/lib/libdwarf/dwarf_memcpy_swap.c
Now uses memcpy to write part of
the line table header, not WRITE_UNALIGNED
as WRITE_UNALIGNED is intended for integers
requiring byte-swapping when the target object
and the runnning libdwarfp do not have the
same endianness. the write changed is a set of
bytes, not an integer at all.
  modified:   src/lib/libdwarfp/dwarf_pro_section.c