Fixes DW202303-044
With that lone backslash  in --search-regex=\\
(or any pattern ending in backslash)
The regex compiler incremented p by one and
then dereferenced and stored the
resulting value. But if the dereference
gave 0 (the NULL byte) this resulted
in unpredictable, but bad, effects
since unknown bytes after the pattern string
would then be referenced as part of the pattern.
Now we notice the improper NUL and issue an error.
	modified:   src/bin/dwarfdump/dd_regex.c