Coverity Scan discovered issues. All are fixed with these changes.
A loop looking for invalid .conf files
was kind of doing the wrong thing.
So it had dead code.
Now doing the right thing and passes
	modified:   src/bin/dwarfdump/dd_dwconf.c

Deleted dead code (constand propagation in
Coverity found it) and replaced with a comment
saying there is no need to check the offset-size.
	modified:   src/lib/libdwarf/dwarf_gnu_index.c

More dead code. Removed it.
	modified:   src/lib/libdwarf/dwarf_rnglists.c