Now using project_build_root() and project_source_root() instead of the deprecated functions build_root() and source_root().
Updated the required version to 0.56 as this
is the version required to support the project*()
commands we now use.

Also added code to look in specific directories
beyond the standard locations
for zlib and libzstd (/opt/local/lib and /usr/local/lib)
so the build can work in more environments.
	modified:   meson.build
	modified:   test/meson.build