Coverityscan found 3 issues.
Here, they were real could lead to a segv.
the issue was the code has options oriented
to testing (probably a mistake to put such in an example)
and there was a dereference of a null pointer
that could occur in a particular scenario. Fixed.
	modified:   src/bin/dwarfexample/simplereader.c

Here the (new, undocumented) option of leaving
either of two pointer inputs (for return values) NUll to
dwarf_get_section_info_by_name was
not coded correctly.  Fixed.
	modified:   src/lib/libdwarf/dwarf_init_finish.c