On linux gdb and valgrind  and libc notice that printf of sequential byte pairs, where the byte-pairs accidentally overlap (!) due to a mistake in this source... is bogus.
MacOS notices too, but it reports "Trace/BPT trap: 5"
which is quite mysterious.
This is a serious bug which has been in place
since support for the DWARF5 .debug_names section
was added.  It crashes the app reading the .debug_info
section reliably.

Of course simplereader is just an example, but
still...needs to be correct

It had been hidden from testing before due to an oversight in
the  debugfissionb directory in regressiontests
was coded. We now prevent corrupt
DWARF with duplicated attributes in a PE file from
showing an error and avoiding
any meaingful test. Just for this case.
So the tests are now fixed too.
	modified:   src/bin/dwarfexample/simplereader.c