Defining (documenting) two existing Segment names. 	modified:   src/lib/libdwarf/dwarf_macho_loader.h
To find corrupted Mach-o objects we are using segment
and section names as a kind of proxy.
We validate segment names against a known list
and we validate section names by ensuring they
are plain ascii.
All these names are defined by an ABI or by the
compiler-writers for the ABI.

This fixes oss fuzz 446726229 (a.k.a DW202509-002)
which was unprepared for this sort of corruption,
starting from the initial handling of Mach-o
objects in 2018.
	modified:   src/lib/libdwarf/dwarf_machoread.c